--- a/.gitignore
+++ b/.gitignore
@@ -2,6 +2,7 @@
 *.o
 *.a
 *.so
+*.out
 
 # Folders
 _obj
@@ -23,6 +24,7 @@ _testmain.go
 *.exe
 *.test
 *.prof
+bench*
 
 # Emacs
 *~