--- a/length_test.go
+++ b/length_test.go
@@ -2,15 +2,16 @@ package jutil
 
 import (
 	"encoding/json"
-	"io/ioutil"
 	"testing"
 	"time"
 
 	"github.com/segmentio/ecs-logs-go"
 )
 
-var (
-	lengthTests = []interface{}{
+const longString = `"Package json implements encoding and decoding of JSON objects as defined in RFC 4627. The mapping between JSON objects and Go values is described in the documentation for the Marshal and Unmarshal functions.")`
+
+func TestLength(t *testing.T) {
+	tests := []interface{}{
 		nil,
 
 		true,
@@ -32,6 +33,9 @@ var (
 
 		json.Number("0"),
 		json.Number("1.2345"),
+
+		time.Now(),
+		12 * time.Hour,
 
 		[]int{},
 		[]int{1, 2, 3},
@@ -80,10 +84,8 @@ var (
 			Message: "Hello World!",
 		},
 	}
-)
 
-func TestLength(t *testing.T) {
-	for _, test := range lengthTests {
+	for _, test := range tests {
 		b, _ := json.Marshal(test)
 
 		if n, err := Length(test); err != nil {
@@ -94,27 +96,148 @@ func TestLength(t *testing.T) {
 	}
 }
 
-func BenchmarkJsonLen(b *testing.B) {
+func benchLength(b *testing.B, v interface{}) {
 	for i := 0; i != b.N; i++ {
-		for _, test := range lengthTests {
-			Length(test)
-		}
+		benchLengthFunc(v)
 	}
 }
 
-func BenchmarkJsonMarshal(b *testing.B) {
-	for i := 0; i != b.N; i++ {
-		for _, test := range lengthTests {
-			json.Marshal(test)
-		}
-	}
+func BenchmarkLengthBoolZero(b *testing.B) {
+	benchLength(b, false)
 }
 
-func BenchmarkJsonMarshalDevNull(b *testing.B) {
-	for i := 0; i != b.N; i++ {
-		for _, test := range lengthTests {
-			e := json.NewEncoder(ioutil.Discard)
-			e.Encode(test)
-		}
-	}
+func BenchmarkLengthBoolNonZero(b *testing.B) {
+	benchLength(b, true)
+}
+
+func BenchmarkLengthIntZero(b *testing.B) {
+	benchLength(b, 0)
+}
+
+func BenchmarkLengthIntNonZero(b *testing.B) {
+	benchLength(b, 1234567890)
+}
+
+func BenchmarkLengthFloatZero(b *testing.B) {
+	benchLength(b, 0.0)
+}
+
+func BenchmarkLengthFloatNonZero(b *testing.B) {
+	benchLength(b, 12345.67890)
+}
+
+func BenchmarkLengthStringZero(b *testing.B) {
+	benchLength(b, "")
+}
+
+func BenchmarkLengthStringNonZero(b *testing.B) {
+	benchLength(b, longString)
+}
+
+func BenchmarkLengthBytesZero(b *testing.B) {
+	benchLength(b, []byte{})
+}
+
+func BenchmarkLengthBytesNonZero(b *testing.B) {
+	benchLength(b, []byte(longString))
+}
+
+func BenchmarkLengthSliceInterfaceZero(b *testing.B) {
+	benchLength(b, []interface{}{})
+}
+
+func BenchmarkLengthSliceInterfaceNonZero(b *testing.B) {
+	benchLength(b, []interface{}{})
+}
+
+func BenchmarkLengthSliceBoolZero(b *testing.B) {
+	benchLength(b, []bool{})
+}
+
+func BenchmarkLengthSliceBoolNonZero(b *testing.B) {
+	benchLength(b, []bool{
+		true, false, true, false, true, false, true, false, true, false, true, false,
+	})
+}
+
+func BenchmarkLengthTimeZero(b *testing.B) {
+	benchLength(b, time.Time{})
+}
+
+func BenchmarkLengthTimeNonZero(b *testing.B) {
+	benchLength(b, time.Now())
+}
+
+func BenchmarkLengthDurationZero(b *testing.B) {
+	benchLength(b, time.Duration(0))
+}
+
+func BenchmarkLengthDurationNonZero(b *testing.B) {
+	benchLength(b, 12*time.Hour)
+}
+
+func BenchmarkLengthMapStringInterfaceZero(b *testing.B) {
+	benchLength(b, map[string]interface{}{})
+}
+
+func BenchmarkLengthMapStringInterfaceNonZero(b *testing.B) {
+	benchLength(b, map[string]interface{}{
+		"0": true,
+		"1": true,
+		"2": true,
+		"3": true,
+		"4": true,
+		"5": true,
+		"6": true,
+		"7": true,
+		"8": true,
+		"9": true,
+	})
+}
+
+func BenchmarkLengthMapStringStringZero(b *testing.B) {
+	benchLength(b, map[string]string{})
+}
+
+func BenchmarkLengthMapStringStringNonZero(b *testing.B) {
+	benchLength(b, map[string]string{
+		"0": "",
+		"1": "",
+		"2": "",
+		"3": "",
+		"4": "",
+		"5": "",
+		"6": "",
+		"7": "",
+		"8": "",
+		"9": "",
+	})
+}
+
+func BenchmarkLengthStructZero(b *testing.B) {
+	benchLength(b, struct{}{})
+}
+
+func BenchmarkLengthStructNonZero(b *testing.B) {
+	benchLength(b, struct {
+		A int
+		B int
+		C int
+	}{1, 2, 3})
+}
+
+func BenchmarkLengthStructOmitEmptyZero(b *testing.B) {
+	benchLength(b, struct {
+		A int `json:",omitempty"`
+		B int `json:",omitempty"`
+		C int `json:",omitempty"`
+	}{})
+}
+
+func BenchmarkLengthStructOmpitemptytNonZero(b *testing.B) {
+	benchLength(b, struct {
+		A int `json:",omitempty"`
+		B int `json:",omitempty"`
+		C int `json:",omitempty"`
+	}{1, 2, 3})
 }