I decoupled the benchmarks and did more accurate comparisons with the standard library, then made a couple of improvements that got the `Length` function from ~2x to ~10x faster than the standard JSON encoder:
```
benchmark                                    old ns/op     new ns/op     delta
BenchmarkLengthBoolZero                      307           17.9          -94.17%
BenchmarkLengthBoolNonZero                   317           17.6          -94.45%
BenchmarkLengthIntZero                       330           21.6          -93.45%
BenchmarkLengthIntNonZero                    350           34.0          -90.29%
BenchmarkLengthFloatZero                     371           86.8          -76.60%
BenchmarkLengthFloatNonZero                  522           250           -52.11%
BenchmarkLengthStringZero                    325           24.7          -92.40%
BenchmarkLengthStringNonZero                 1052          933           -11.31%
BenchmarkLengthBytesZero                     377           22.3          -94.08%
BenchmarkLengthBytesNonZero                  1145          22.5          -98.03%
BenchmarkLengthSliceInterfaceZero            326           25.8          -92.09%
BenchmarkLengthSliceInterfaceNonZero         337           25.6          -92.40%
BenchmarkLengthSliceBoolZero                 331           62.6          -81.09%
BenchmarkLengthSliceBoolNonZero              943           846           -10.29%
BenchmarkLengthTimeZero                      1090          531           -51.28%
BenchmarkLengthTimeNonZero                   1254          574           -54.23%
BenchmarkLengthDurationZero                  333           68.4          -79.46%
BenchmarkLengthDurationNonZero               361           91.6          -74.63%
BenchmarkLengthMapStringInterfaceZero        507           19.4          -96.17%
BenchmarkLengthMapStringInterfaceNonZero     5783          843           -85.42%
BenchmarkLengthMapStringStringZero           504           165           -67.26%
BenchmarkLengthMapStringStringNonZero        4848          3649          -24.73%
BenchmarkLengthStructZero                    326           99.1          -69.60%
BenchmarkLengthStructNonZero                 668           415           -37.87%
BenchmarkLengthStructOmitEmptyZero           411           177           -56.93%
BenchmarkLengthStructOmpitemptytNonZero      700           430           -38.57%
```